+++ src/components/POLine/ChangeInstanceConnection/ConsortiumRelatedItemsList/index.js
@@ -1,0 +1,1 @@
+export { ConsortiumRelatedItemsList } from './ConsortiumRelatedItemsList';