+++ src/components/POLine/ChangeInstanceConnection/hooks/useRelatedItemsMCL/index.js
@@ -1,0 +1,1 @@
+export { useRelatedItemsMCL } from './useRelatedItemsMCL';