+++ src/components/POLine/ChangeInstanceConnection/utils/index.js
@@ -1,1 +1,2 @@
+export { checkIfSubmitActionProhibited } from './checkIfSubmitActionProhibited';
 export { getUpdateHoldingsOptions } from './getUpdateHoldingsOptions';